+++ lib/reports/html_report.py
@@ -18,9 +18,10 @@
 
 import json
 import os
-from mako.template import Template
 
 from lib.reports import *
+from lib.utils import FileUtils
+from thirdparty.mako.template import Template
 
 
 class HTMLReport(FileBaseReport):
@@ -28,21 +29,36 @@ class HTMLReport(FileBaseReport):
         template_file = os.path.dirname(os.path.realpath(__file__)) + '/templates/html_report_template.html'
         mytemplate = Template(filename=template_file)
 
+        metadata = {
+            "command": " ".join(sys.argv),
+            "date": time.ctime()
+        }
         results = []
         for entry in self.entries:
             for e in entry.results:
                 headerName = "{0}://{1}:{2}/{3}".format(
                     entry.protocol, entry.host, entry.port, entry.basePath
                 )
+
+                statusColorClass = ''
+                if e.status >= 200 and e.status <= 299:
+                    statusColorClass = 'text-success'
+                elif e.status >= 300 and e.status <= 399:
+                    statusColorClass = 'text-warning'
+                elif e.status >= 400 and e.status <= 599:
+                    statusColorClass = 'text-danger'
+
                 results.append({
                     "url": headerName + e.path,
                     "path": e.path,
                     "status": e.status,
-                    "contentLength": e.getContentLength(),
+                    "statusColorClass": statusColorClass,
+                    "contentLength": FileUtils.size_human(e.getContentLength()),
+                    "contentType": e.response.headers.get("content-type"),
                     "redirect": e.response.redirect
                 })
 
-        return mytemplate.render(results=json.dumps(results))
+        return mytemplate.render(metadata=metadata, results=json.dumps(results))
 
     def save(self):
         self.file.seek(0)