+++ thirdparty/ntlm_auth/exceptions.py
@@ -1,0 +1,6 @@
+# Copyright: (c) 2018, Jordan Borean (@jborean93) <jborean93@gmail.com>
+# MIT License (see LICENSE or https://opensource.org/licenses/MIT)
+
+
+class NoAuthContextError(Exception):
+    pass